Update recommender results
Adds two new datasets with recommendations:
(i) recommendations.csv
(ii) recommendations_bias_new_york.csv

Both have continious values for the improvement (RMSE), and also
(ii) has some additional bias so that NY populations in specific
are presented with improvement inversly to their age.

The purpose of (ii) along with the pictures added in this commit,
is to show-case a scenario that bridges error distribution use case
and discriminatory associations.